+++ CHANGELOG.md
@@ -201,3 +201,4 @@
   - Scare Spiders and Scare Zombies are no longer Mutation genes, and don't require Scare Creepers or Scare Skeletons
 - Disabling Genes now uses the Gene tag `#geneticsresequenced:disabled`
 - Fixed an issue where the Photosynthesis Gene would stop working when enabled instead of when not enabled
+- Added Reaching Gene, which allows you to reach 1.5 times further